+++ Assets/01.Scripts/System/Item/Database/IItemData.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public abstract class IItemData : ScriptableObject
 {
-    public abstract void RegisterItemData(Dictionary<ItemRarity, ItemDatabaseTable> database, int weight);
+    public abstract void RegisterItemData(Dictionary<ItemRarity, ItemDatabaseTable> database, float weight);
     public abstract ItemSO PickRandom();
 }
